src/main/java/com/computeralchemist/domain/components/disk/DiskParameters.java
package com.computeralchemist.domain.components.disk;

/**
 * @Author
 * Karol Meksuła
 * 07-04-2018
 * */

public class DiskParameters {
    private DriveType type;
    private String format;
    private int capacity;
    private double readSpeed;
    private double writeSpeed;

    public DriveType getType() {
        return type;
    }

    public void setType(DriveType type) {
        this.type = type;
    }

    public String getFormat() {
        return format;
    }

    public void setFormat(String format) {
        this.format = format;
    }

    public int getCapacity() {
        return capacity;
    }

    public void setCapacity(int capacity) {
        this.capacity = capacity;
    }

    public double getReadSpeed() {
        return readSpeed;
    }

    public void setReadSpeed(double readSpeed) {
        this.readSpeed = readSpeed;
    }

    public double getWriteSpeed() {
        return writeSpeed;
    }

    public void setWriteSpeed(double writeSpeed) {
        this.writeSpeed = writeSpeed;
    }
}
